requirements.txt
python-binance
pandas
numpy
supabase
requests